+++ op2/src/openmp4/op_openmp4_decl.cpp
@@ -9,9 +9,9 @@
 
 void cutilDeviceInit(int argc, char **argv);
 
-void op_mvHostToDevice(void **map, int size);
+void op_mvHostToDevice(void **map, size_t size);
 
-void op_cpHostToDevice(void **data_d, void **data_h, int size);
+void op_cpHostToDevice(void **data_d, void **data_h, size_t size);
 
 void op_cuda_exit();
 
@@ -60,11 +60,11 @@ op_dat op_decl_dat_char(op_set set, int dim, char const *type, int size,
       }
     }
     op_cpHostToDevice((void **)&(dat->data_d), (void **)&(temp_data),
-                      dat->size * set->size);
+                      (size_t)dat->size * set->size);
     free(temp_data);
   } else {
     op_cpHostToDevice((void **)&(dat->data_d), (void **)&(dat->data),
-                      dat->size * set->size);
+                      (size_t)dat->size * set->size);
   }
 
   return dat;
@@ -116,11 +116,11 @@ op_dat op_decl_dat_temp_char(op_set set, int dim, char const *type, int size,
       }
     }
     op_cpHostToDevice((void **)&(dat->data_d), (void **)&(temp_data),
-                      dat->size * set->size);
+                      (size_t)dat->size * set->size);
     free(temp_data);
   } else {
     op_cpHostToDevice((void **)&(dat->data_d), (void **)&(dat->data),
-                      dat->size * set->size);
+                      (size_t)dat->size * set->size);
   }
 
   return dat;
@@ -149,7 +149,7 @@ op_map op_decl_map(op_set from, op_set to, int dim, int *imap,
   }
 
   op_cpHostToDevice((void **)&(map->map_d), (void **)&(temp_map),
-                    map->dim * set_size * sizeof(int));
+                    (size_t)map->dim * set_size * sizeof(int));
   free(temp_map);
   return map;
 }
